requirements.txt
langgraph==0.2.53
langchain==0.3.14
langchain-openai==0.2.14
langchain-anthropic==0.3.3
fastapi==0.115.6
uvicorn==0.34.0
requests==2.32.3
pydantic==2.10.6
python-dotenv==1.0.1
